Added direct pass-in and pass-out support
ode_nanowire() can now return delta_N, delta_P directly (without writing to a file)
propagatingPL() can now accept rad_rec as a parameter (without reading from a file)

Because both of these may be called hundreds of times in a bayesim script but their results are only intermediates, it is sometimes not practical to save all data. With these changes, the option to NOT save all the data now exists.